requirements/base.txt
asgiref==3.7.2
django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
django_celery_beat==2.5.0
flower==2.0.1
flake8==7.0.0
django-split-settings==1.2.0
django-environ==0.11.2
django-cors-headers==3.14.0
python-dotenv==1.0.1